+++ js.bolt.surge/flake.nix
@@ -2,8 +2,7 @@
   description = "a powerful online email application";
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs/nixos-unstable";
-    nixpkgs-biome.url = "github:NixOS/nixpkgs/0d534853a55b5d02a4ababa1d71921ce8f0aee4c"; # 1.9.4
-    nixpkgs-node.url = "github:NixOS/nixpkgs/0cb2fd7c59fed0cd82ef858cbcbdb552b9a33465"; # 22.5.1
+    nixpkgs-biome.url = "github:NixOS/nixpkgs/c2d7b8bfc494f234d191322a7c387a9ff67e1786"; # 2.1.2
     flake-utils.url = "github:numtide/flake-utils";
   };
   outputs =
@@ -11,7 +10,6 @@
       flake-utils,
       nixpkgs,
       nixpkgs-biome,
-      nixpkgs-node,
       ...
     }:
     flake-utils.lib.eachDefaultSystem (
@@ -23,21 +21,19 @@
         biomepkgs = import nixpkgs-biome {
           inherit system;
         };
-        nodepkgs = import nixpkgs-node {
-          inherit system;
-        };
+        # https://github.com/slackapi/slack-cli
         slackcli = pkgs.stdenv.mkDerivation {
           name = "slackcli";
           src =
             if pkgs.stdenv.isDarwin then
               pkgs.fetchurl {
-                url = "https://downloads.slack-edge.com/slack-cli/slack_cli_3.0.0_macOS_64-bit.tar.gz";
-                sha256 = "0xbld1a01354zvh5j8l8cgh4b8lgwrk5ngq04i1vdrfjp981a697";
+                url = "https://downloads.slack-edge.com/slack-cli/slack_cli_3.5.1_macOS_64-bit.tar.gz";
+                sha256 = "0abavvsjp1mi6s9wvbishswwr5jh7s48b71nnafvfhhhv261x56i";
               }
             else
               pkgs.fetchurl {
-                url = "https://downloads.slack-edge.com/slack-cli/slack_cli_3.0.0_linux_64-bit.tar.gz";
-                sha256 = "1lrapyy3lw9gg919nfim6vnnwswmc7kyfr8ywmy3lq5py8gdmgb3";
+                url = "https://downloads.slack-edge.com/slack-cli/slack_cli_3.5.1_linux_64-bit.tar.gz";
+                sha256 = "15h5zsn1h6lz5q0qsy9qhjphrf516npqhpixy9w7f1ill0gk1xn2";
               };
           unpackPhase = "tar -xzf $src";
           installPhase = ''
@@ -52,9 +48,10 @@
             pkgs.bash # https://git.savannah.gnu.org/cgit/bash.git
             biomepkgs.biome # https://github.com/biomejs/biome
             pkgs.heroku # https://github.com/heroku/cli
-            nodepkgs.nodejs_22 # https://github.com/nodejs/node
+            pkgs.jq # https://github.com/jqlang/jq
+            pkgs.nodejs_22 # https://github.com/nodejs/node
             pkgs.opentofu # https://github.com/opentofu/opentofu
-            slackcli # https://tools.slack.dev/slack-cli
+            slackcli
           ];
           shellHook = ''
             export SLACK_CONFIG_DIR="$HOME/.config/slack"