+++ deno.sdk.chanl/CHANGELOG.md
@@ -7,6 +7,7 @@ versioned with [calendar versioning][calver].
 
 ## Changes
 
+- build: update to the latest runtime version and developer tools 2025-07-26
 - feat: create ephemeral channels for quick testing and archival 2024-07-12
 
 [calver]: https://calver.org