+++ js.bolt.surge/CHANGELOG.md
@@ -7,6 +7,7 @@ versioned with [calendar versioning][calver].
 
 ## Changes
 
+- build: update node to the latest version to match flake 2025-07-26
 - ci: separate production environments during release run 2025-05-18
 - build: guard against important manifest object override 2025-03-15
 - test: track coverage of unit tests for updating insight 2025-03-15